Move GoCD eventemitter Init
