+++ src/brain/gocdNoDeploysAlert/index.ts
@@ -1,6 +1,5 @@
 import { KnownBlock } from '@slack/types';
 
-import { gocdevents } from '@/api/gocdevents';
 import * as slackblocks from '@/blocks/slackBlocks';
 import {
   FEED_DEV_INFRA_CHANNEL_ID,
@@ -8,6 +7,7 @@ import {
   GOCD_SENTRYIO_BE_PIPELINE_NAME,
   GOCD_SENTRYIO_FE_PIPELINE_NAME,
 } from '@/config';
+import { gocdevents } from '@/init/gocd';
 import { bolt } from '@/init/slack';
 import {
   DBGoCDDeployment,
